move instconfig
