+++ pypit/arspecobj.py
@@ -6,11 +6,13 @@ import numpy as np
 import copy
 
 from pypit import armsgs
+from pypit import arparse as settings
 
 # Logging
 msgs = armsgs.get_logger()
 
 from pypit import ardebug as debugger
+
 
 class SpecObjExp(object):
     """Class to handle object spectra from a single exposure
@@ -74,6 +76,7 @@ class SpecObjExp(object):
         self.boxcar = {}   # Boxcar extraction 'wave', 'counts', 'var', 'sky', 'mask', 'flam', 'flam_var'
         self.optimal = {}  # Optimal extraction 'wave', 'counts', 'var', 'sky', 'mask', 'flam', 'flam_var'
         #
+
     def check_trace(self, trace, toler=1.):
         """Check that the input trace matches the defined specobjexp
 
@@ -287,3 +290,58 @@ def get_objid(slf, det, islit, iobj, trc_img, ypos=0.5):
     objid= int(np.round(xobj*1e3))
     # Return
     return objid, xobj
+
+
+def instconfig(det, scidx, fitsdict):
+    """ Returns a unique config string for the current slf
+
+    Parameters
+    ----------
+    det : int
+    scidx : int
+       Exposure index (max=9999)
+    fitsdict : dict
+    """
+
+    from collections import OrderedDict
+    config_dict = OrderedDict()
+    config_dict['S'] = 'slitwid'
+    config_dict['D'] = 'dichroic'
+    config_dict['G'] = 'dispname'
+    config_dict['T'] = 'dispangle'
+    #
+    config = ''
+    for key in config_dict.keys():
+        try:
+            comp = str(fitsdict[config_dict[key]][scidx])
+        except KeyError:
+            comp = '0'
+        #
+        val = ''
+        for s in comp:
+            if s.isdigit():
+                val += s
+        config = config + key+'{:s}-'.format(val)
+    # Binning
+    try:
+        binning = settings.spect['det'][det-1]['binning']
+    except KeyError:
+        msgs.warn("Assuming 1x1 binning for your detector")
+        binning = '1x1'
+    val = ''
+    for s in binning:
+        if s.isdigit():
+            val = val + s
+    config += 'B{:s}'.format(val)
+    # Return
+    return config
+
+    """
+    msgs.warn("Flat indexing needs to be improved in arsort.setup")
+    fidx = slf._name_flat.index(slf._mspixelflat_name)
+    if fidx > 9:
+        msgs.error("Not ready for that many flats!")
+    aidx = slf._name_flat.index(slf._mspixelflat_name)
+    setup = 10*(aidx+1) + fidx
+    return setup
+    """
